requirements.txt
click==7.1.1
enum-compat==0.0.3
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
paho-mqtt==1.5.0
python-dotenv==0.12.0
six==1.14.0
Werkzeug==1.0.1
yeelight==0.5.1
